\name{analysis.window}
\alias{analysis.window}
\title{Analysis.window}
\description{
Auxiliar function of genArise GUI, in this window you can apply
operations to original data.
}
\usage{
analysis.window(texto, follow.wizard = FALSE, envir)
}
\arguments{
  \item{texto}{Historial project string}
  \item{follow.wizard}{Boolean value, if this argument is TRUE, an data
    analysis are performed}
  \item{envir}{Environment where are the project variables}
}
\value{
  tkwidget
  }
\keyword{math}