\name{classes}
\alias{classes}
\title{Classes Defined by this Package}
\description{
This package defines the following data classes.
\describe{
\item{ \code{\link[genArise:Spot]{Spot}} }{
	A class used to store spot data with the following attributes:
	Cy3, Cy5, BgCy3, BgCy5, Ids as they are read by \code{read.spot}
      or obtained from a function that return a spot object.}
\item{ \code{\link[limma:DataSet]{DataSet}} }{
  A class used to store spot data with the following attributes:
  Cy3, Cy5, Ids, Z-score.}
}}
\keyword{documentation}
