\name{create.project}
\alias{create.project}

\title{Create directorie for the project and its results and graphics}
\description{
Auxiliar function for genAriseGUI. Create the directory's hierarchy of
the project.
}
\usage{
create.project(project.name, results.file = "Results",
               graphics.file = "Graphics")
}
\arguments{
  \item{project.name}{Project directory name.}
  \item{results.file}{Filename of the project result.}
  \item{graphics.file}{Filename of the project graphics.}
}
\keyword{arith}
