\name{genArise.init}
\alias{genArise.init}
\title{genArise.init}
\description{
Auxiliar function of genArise GUI, this function show a principal menu
of genAriseGUI
}
\usage{
genArise.init(envir)
}
\arguments{
  \item{envir}{Environment where are the project variables}
}
\value{
  tkwidget
}
\keyword{math}