\name{graphic.choose}
\alias{graphic.choose}
\title{Graphic choose}
\description{
  This function show the plot of an spot sobject. This plot are identify
  with the graphic.type.value
}
\usage{
graphic.choose(spot.object, graphic.type)
}
\arguments{
  \item{spot.object}{An object ob Spot class}
  \item{graphic.type}{representative integer of type graphic will be plot}
}
\value{
  Plot device
}
\keyword{math}