\name{ma.plot}
\alias{ma.plot}
\title{Data Visualization: M vs. A plot}
\description{
This function allows to plot \bold{M -vs- A} in spot.
}
\usage{
ma.plot(mySpot, col = "green")
}
\arguments{
  \item{mySpot}{Spot for one microarray.}
  \item{col}{color of points in graphic}
}
\examples{
data(Simon)
## plot the signals for spot.
ma.plot(Simon)
}
\keyword{aplot}
\keyword{arith}
