\name{old.project}
\alias{old.project}
\title{Open previous project}
\description{
  Show the information that was obtained from the analysis of a
  previous project. This is just an auxiliar function for genAriseGUI,
  and can not be used in the command line.
}
\usage{
old.project(project.name,envir, parent)
}
\arguments{
  \item{project.name}{path of project file (PRJ)}
  \item{envir}{Environment where are the genArise variables}
  \item{parent}{The parent widget}
}
\value{
  tkwidget
}
\keyword{file}