\name{reset.history}
\alias{reset.history}
\title{Reset the prj history file}
\description{
Clean all the operations saved in the prj history file.
}
\usage{
reset.history(history.file, text)
}
\arguments{
  \item{history.file}{The name of the prj history file.}
  \item{text}{The new content of the prj history file.}
}
\value{
  The history file without operations.
}
\keyword{file}
