\name{set.grid.properties}
\alias{set.grid.properties}
\title{set.grid.properties}
\description{
Auxiliary function for genAriseGUI
}
\usage{
set.grid.properties(envir, name, nr, nc, nmr, nmc)
}
\arguments{
  \item{envir}{Environment where the variables are stored}
  \item{name}{The name of the experiment}
  \item{nr}{Total rows in the array (each row represent a spot)}
  \item{nc}{Total columns in the array}
  \item{nmr}{Total of meta-rows}
  \item{nmc}{Total of meta-columns}
}
\keyword{IO}