\name{set.path.project}
\alias{set.path.project}
\title{set.path.project}
\description{
Auxiliar function for genAriseGUI
}
\usage{
set.path.project(path, results.file, graphics.file, envir)
}
\arguments{
  \item{path}{Project path value}
  \item{results.file}{Name of directory where results file will be}
  \item{graphics.file}{Name of directory where pdf graphics will be}
  \item{envir}{Environment where are the experiment variables}
}
\keyword{arith}