\name{single.norm}
\alias{single.norm}
\title{Swap from Files}
\description{
Read both files, but only extract the interested columns and create a Spot object.
}
\usage{
single.norm(envir)
}
\arguments{
\item{envir}{Environment where are the genArise variables.}
}
\seealso{
\code{\link{write.spot}}.
}
\keyword{file}
