\name{swap.select}
\alias{swap.select}
\title{Dye swap files selector}
\description{
  This is just an auxiliar function for genAriseGUI, and can not be used
  in the command line.
}
\usage{
swap.select(envir)
}
\arguments{
  \item{envir}{Environment where are the genArise variables}
}
\value{
  tkwidget
}
\keyword{file}