\name{create.project}
\alias{create.project}
\title{create.project}
\description{
Auxiliar function for genAriseGUI
}
\usage{
create.project(project.name, results.file = "Results",
graphics.file = "Graphics")
}
\arguments{
  \item{project.name}{}
  \item{results.file}{}
  \item{graphics.file}{}
}
\keyword{arith}